requirements.txt
google-generativeai
fastapi
uvicorn
streamlit